+++ tests/AST/BinaryOperators/ASTVisitor_BinaryOperators_Relational_Tests.cpp
@@ -14,35 +14,10 @@ static const char *const fakeSourceFilePath = "input.cc";
 
 static FilePathFilter nullPathFilter;
 
-TEST(ASTVisitorTest, binaryAddOperator) {
+TEST(ASTVisitorTest, basicConditional) {
   const char *const binaryOperator = R"(///
 int add(int a, int b) {
-  return a + b;
-}
-)";
-
-  ASTStorage storage("", "");
-
-  std::unique_ptr<clang::ASTUnit> astUnit(
-      clang::tooling::buildASTFromCode(binaryOperator, fakeSourceFilePath));
-
-  ASTMutations astMutations;
-
-  ThreadSafeASTUnit threadSafeAstUnit(std::move(astUnit));
-  ASTVisitor astVisitor(threadSafeAstUnit, astMutations,
-                        nullPathFilter,
-                        mull::TraverseMask::BINARY_OP);
-
-  astVisitor.traverse();
-
-  ASSERT_EQ(astMutations.count(), 1U);
-  ASSERT_TRUE(astMutations.locationExists(fakeSourceFilePath, 3, 12));
-}
-
-TEST(ASTVisitorTest, binarySubOperator) {
-  const char *const binaryOperator = R"(///
-int add(int a, int b) {
-  return a - b;
+  return a < b;
 }
 )";
